+++ cm_backtrace/cm_backtrace.h
@@ -40,7 +40,7 @@ void cm_backtrace_init(const char *firmware_name, const char *hardware_ver, cons
 void cm_backtrace_firmware_info(void);
 size_t cm_backtrace_call_stack(uint32_t *buffer, size_t size, uint32_t sp);
 
-void cm_backtrace_assert(const char *expr, const char *file, size_t line);
+void cm_backtrace_assert(const char *expr, const char *file, unsigned line);
 void cm_backtrace_fault(uint32_t fault_handler_lr, uint32_t fault_handler_sp);
 
 #ifdef  __cplusplus
